+++ VisualCard.Calendar/Parts/CalendarEvent.cs
@@ -35,7 +35,7 @@ namespace VisualCard.Calendar.Parts
         internal readonly List<CalendarAlarm> alarms = [];
         private readonly Dictionary<CalendarPartsArrayEnum, List<BaseCalendarPartInfo>> partsArray = [];
         private readonly Dictionary<CalendarStringsEnum, string> strings = [];
-        private readonly Dictionary<CalendarIntegersEnum, int> integers = [];
+        private readonly Dictionary<CalendarIntegersEnum, double> integers = [];
 
         /// <summary>
         /// Alarm list
@@ -125,7 +125,7 @@ namespace VisualCard.Calendar.Parts
             hashCode = hashCode * -1521134295 + EqualityComparer<List<CalendarAlarm>>.Default.GetHashCode(alarms);
             hashCode = hashCode * -1521134295 + EqualityComparer<Dictionary<CalendarPartsArrayEnum, List<BaseCalendarPartInfo>>>.Default.GetHashCode(partsArray);
             hashCode = hashCode * -1521134295 + EqualityComparer<Dictionary<CalendarStringsEnum, string>>.Default.GetHashCode(strings);
-            hashCode = hashCode * -1521134295 + EqualityComparer<Dictionary<CalendarIntegersEnum, int>>.Default.GetHashCode(integers);
+            hashCode = hashCode * -1521134295 + EqualityComparer<Dictionary<CalendarIntegersEnum, double>>.Default.GetHashCode(integers);
             return hashCode;
         }
 
@@ -143,7 +143,7 @@ namespace VisualCard.Calendar.Parts
         internal new void SetString(CalendarStringsEnum key, string value) =>
             SetString(key, value, strings);
 
-        internal new void SetInteger(CalendarIntegersEnum key, int value) =>
+        internal new void SetInteger(CalendarIntegersEnum key, double value) =>
             SetInteger(key, value, integers);
 
         internal CalendarEvent(Version version) :
